First pass of design complete
This was a major update which mainly focused on fleshing out the concept, design and function.

Improved the html structure, added several class selectors, a page navigation system and split the content into paragraphs.

Completed basic CSS with the aim of putting most styling options behind class selectors.

Created a new folder for images and added a logo image.