Added a main header and a top nav
Added a page header, the fonts I'm going to be using, made the size of the page the middle 75% so you have some (hopefully) nice white space on either side.

Added a top bar nav which sits next to the header, these will be internal links when I get around to it.