HTML/CSS Successfully Using First-child Pseudo Class
Moved the featured code block into the same code block as the tv-shows/movie/web-series sections.

Changed the Css to using first -child and not nth-of-type.

Placed a border around the featured sections of the web-page to ensure changes happened to the correct div.

Used flex-wrap and flex-direction row to ensure any additional content with each section wraps below the featured section.

Used the flex property on the favourites class to ensure each section have equal sizing and overwrite the size of the featured section so it takes up the whole width of its container.